# swagger.yml
openapi: 3.0.2
info:
  title: API Ercole 2.0
  version: 1.0.0
  license:
    name: GNU General Public License v3.0 or later
    url: 'https://www.gnu.org/licenses/gpl-3.0-standalone.html'
  description: API used to talk to ercole services
  contact:
    name: Issues support
    url: 'https://github.com/ercole-io/ercole'
servers:
  - description: Api service on dev.ercole.io
    url: 'https://dev.ercole.io/api'
  - description: Data service on dev.ercole.io
    url: 'https://dev.ercole.io/data'
  - description: Alert service on dev.ercole.io
    url: 'https://dev.ercole.io/alert'
  - description: Repo service on dev.ercole.io
    url: 'https://dev.ercole.io/repo'
  - description: localhost address
    url: 'http://localhost:11111'
...
components:
  securitySchemes:
    basicAuthDataService:
      type: http
      scheme: basic
      description: access to data-service APIs
    basicAuthAlertService:
      type: http
      scheme: basic
      description: access to alert-service APIs
    basicAuthApiService:
      type: http
      scheme: basic
      description: access to api-service APIs
    tokenAuthApiService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: access using the access token
  responses:
    error:
      description: Contains some informations about a error
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required:
              - Error
              - ErrorDescription
              - SourceFilename
              - LineNumber
            properties:
              Error:
                type: string
              ErrorDescription:
                type: string
              SourceFilename:
                type: string
              LineNumber:
                type: integer
          examples:
            Test1:
              value:
                Error: Invalid  status
                ErrorDescription: invalid status
                SourceFilename: /home/travis/go/src/github.com/ercole-io/ercole/api-service/controller/alerts_api.go
                LineNumber: 65
  schemas:
...
    CellDisk:
      type: object
      properties:
        Name:
          type: string
        Status:
          type: string
        ErrCount:
          type: string
        UsedPerc:
          type: string
      required:
        - Name
        - Status
        - ErrCount
        - UsedPerc
    ExadataDev:
      type: object
      properties:
        Hostname:
          type: string
        ServerType:
          type: string
          nullable: true
        Model:
          type: string
        ExaSwVersion:
          type: string
        CPUEnabled:
          type: string
          nullable: true
        Memory:
          type: string
          nullable: true
        Status:
          type: string
          nullable: true
        PowerCount:
          type: string
          nullable: true
        PowerStatus:
          type: string
          nullable: true
        FanCount:
          type: string
          nullable: true
        FanStatus:
          type: string
          nullable: true
        TempActual:
          type: string
          nullable: true
        TempStatus:
          type: string
          nullable: true
        CellsrvService:
          type: string
          nullable: true
        MsService:
          type: string
          nullable: true
        RsService:
          type: string
          nullable: true
        FlashcacheMode:
          type: string
          nullable: true
        CellDisks:
          nullable: true
          type: array
          items:
            $ref: '#/components/schemas/CellDisk'
      required:
        - Hostname
        - Model
        - ExaSwVersion
...
    PageMetadata:
      type: object
      required:
        - Empty
        - First
        - Last
        - Number
        - Size
        - TotalElements
        - TotalPages
      properties:
        Empty:
          type: boolean
        First:
          type: boolean
        Last:
          type: boolean
        Number:
          type: integer
        Size:
          type: integer
        TotalElements:
          type: integer
        TotalPages:
          type: integer
...
    ExadataSummary:
      type: object
      required:
        - Hostname
        - Location
        - Environment
        - CreatedAt
        - DBServers
        - StorageServers
        - IBSwitches
      properties:
        Hostname:
          type: string
        Location:
          type: string
        Environment:
          type: string
        CreatedAt:
          type: string
        DBServers:
          $ref: '#/components/schemas/ExadataDev'
        StorageServers:
          $ref: '#/components/schemas/ExadataDev'
        IBSwitches:
          $ref: '#/components/schemas/ExadataDev'
...
  parameters:
    search:
      in: query
      name: search
      required: false
      description: filter the result using search terms
      allowEmptyValue: true
      example: foobar
      schema:
        type: string
        readOnly: true
        example: foobar
    sort-by:
      in: query
      name: sort-by
      required: false
      description: sort the result by a field
      allowEmptyValue: true
      example: foobar
      schema:
        type: string
        readOnly: true
        example: CreatedAt
    sort-desc:
      schema:
        type: boolean
      in: query
      name: sort-desc
      description: |
        Sort the field descendently instead of ascendently
      allowEmptyValue: true
    page:
      schema:
        type: integer
      in: query
      name: page
      description: Number of the page
      allowEmptyValue: true
    size:
      schema:
        type: integer
      in: query
      name: size
      description: Size of the page
      allowEmptyValue: true
    location:
      schema:
        type: string
        example: Italy
      in: query
      name: location
      description: Filter by location
      allowEmptyValue: true
    environment:
      schema:
        type: string
        example: TST
      in: query
      name: environment
      description: Filter by environment
      allowEmptyValue: true
    older-than:
      schema:
        type: string
        format: date-time
        example: '2020-05-05T08:23:40+00:00'
      in: query
      name: older-than
      description: Filter until the date
      allowEmptyValue: true
...
paths:
...
  /hosts/technologies/oracle/exadata:
    get:
      summary: Search a list of exadata
      tags:
        - api-service
        - fe-user
        - read
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/ExadataSummary'
                  - type: object
                    properties:
                      Content:
                        type: array
                        items:
                          $ref: '#/components/schemas/ExadataSummary'
                      Metadata:
                        $ref: '#/components/schemas/PageMetadata'
                    required:
                      - Content
                      - Metadata
      operationId: SearchOracleExadata
      description: |-
        Search exadata data using the filters in the request.
        Return content type could be JSON or XLSX.
      parameters:
        - $ref: '#/components/parameters/search'
        - $ref: '#/components/parameters/sort-by'
        - $ref: '#/components/parameters/sort-desc'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/location'
        - $ref: '#/components/parameters/environment'
        - $ref: '#/components/parameters/older-than'
...
  /hosts/technologies/oracle/exadata/total-memory-size:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: GetTotalOracleExadataMemorySizeStats
      summary: Get the total size of the memory of the exadata
      description: Get the total size of the memory of the exadata
      parameters:
        - $ref: '#/components/parameters/location'
        - $ref: '#/components/parameters/environment'
        - $ref: '#/components/parameters/older-than'
      responses:
        '200':
          description: statistics
          content:
            text/plain:
              schema:
                type: number
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  /hosts/technologies/oracle/exadata/total-cpu:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: GetTotalOracleExadataCPUStats
      summary: Get the total cpu of the exadata
      description: Get the total cpu of the exadata
      parameters:
        - $ref: '#/components/parameters/location'
        - $ref: '#/components/parameters/environment'
        - $ref: '#/components/parameters/older-than'
      responses:
        '200':
          description: statistics
          content:
            application/json:
              schema:
                type: object
                required:
                  - Running
                  - Total
                properties:
                  Running: 
                    type: number
                  Total: 
                    type: number
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error' 
  /hosts/technologies/oracle/exadata/average-storage-usage:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: GetAverageOracleExadataStorageUsageStats
      summary: Get the average usage of the storage of the exadata
      description: Get the average usage of the storage of the exadata
      parameters:
        - $ref: '#/components/parameters/location'
        - $ref: '#/components/parameters/environment'
        - $ref: '#/components/parameters/older-than'
      responses:
        '200':
          description: statistics
          content:
            text/plain:
              schema:
                type: number
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  /hosts/technologies/oracle/exadata/storage-error-count-status:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: GetOracleExadataStorageErrorCountStatusStats
      summary: Get the statistics about the ErrorCount status of the storage of the exadata
      description: Get the statistics about the ErrorCount status of the storage of the  exadata
      parameters:
        - $ref: '#/components/parameters/location'
        - $ref: '#/components/parameters/environment'
        - $ref: '#/components/parameters/older-than'
      responses:
        '200':
          description: statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    Failing:
                      type: boolean
                    Count:
                      type: integer
                  required:
                    - Failing
                    - Count
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  /hosts/technologies/oracle/exadata/patch-status:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: GetOracleExadataPatchStatusStats
      summary: Get the statistics about the patch status of the exadata
      description: Get the statistics about the patch status of the exadata
      parameters:
        - $ref: '#/components/parameters/location'
        - $ref: '#/components/parameters/environment'
        - $ref: '#/components/parameters/older-than'
      responses:
        '200':
          description: statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    Status:
                      type: boolean
                    Count:
                      type: integer
                  required:
                    - Status
                    - Count
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
...
security:
  - basicAuthApiService: []
  - tokenAuthApiService: []